moving remaining beacon tests to pytest.
